+++ sentry/src/test/java/io/sentry/SentryWrapperTest.kt
@@ -195,4 +195,85 @@ class SentryWrapperTest {
             assertEquals(threadedScopes, Sentry.getCurrentScopes())
         }.get()
     }
+
+    @Test
+    fun `wrapped runnable isolates Scopes`() {
+        val capturedEvents = mutableListOf<SentryEvent>()
+
+        initForTest {
+            it.dsn = dsn
+            it.beforeSend = SentryOptions.BeforeSendCallback { event, hint ->
+                capturedEvents.add(event)
+                event
+            }
+        }
+
+        Sentry.addBreadcrumb("MyOriginalBreadcrumbBefore")
+        Sentry.captureMessage("OriginalMessageBefore")
+        println(Thread.currentThread().name)
+
+        val future1 = executor.submit(
+            SentryWrapper.wrapRunnable {
+                Thread.sleep(20)
+                Sentry.addBreadcrumb("MyClonedBreadcrumb")
+                Sentry.captureMessage("ClonedMessage")
+                "Result 1"
+            }
+        )
+
+        val future2 = executor.submit(
+            SentryWrapper.wrapRunnable {
+                Thread.sleep(10)
+                Sentry.addBreadcrumb("MyClonedBreadcrumb2")
+                Sentry.captureMessage("ClonedMessage2")
+                "Result 2"
+            }
+        )
+
+        Sentry.addBreadcrumb("MyOriginalBreadcrumb")
+        Sentry.captureMessage("OriginalMessage")
+
+        future1.get()
+        future2.get()
+
+        val mainEvent = capturedEvents.firstOrNull { it.message?.formatted == "OriginalMessage" }
+        val clonedEvent = capturedEvents.firstOrNull { it.message?.formatted == "ClonedMessage" }
+        val clonedEvent2 = capturedEvents.firstOrNull { it.message?.formatted == "ClonedMessage2" }
+
+        assertEquals(2, mainEvent?.breadcrumbs?.size)
+        assertEquals(2, clonedEvent?.breadcrumbs?.size)
+        assertEquals(2, clonedEvent2?.breadcrumbs?.size)
+    }
+
+    @Test
+    fun `scopes is reset to state within the thread after isolated runnable is done`() {
+        initForTest {
+            it.dsn = dsn
+        }
+
+        val mainScopes = Sentry.getCurrentScopes()
+        val threadedScopes = Sentry.getCurrentScopes().forkedCurrentScope("test")
+
+        executor.submit {
+            Sentry.setCurrentScopes(threadedScopes)
+        }.get()
+
+        assertEquals(mainScopes, Sentry.getCurrentScopes())
+
+        val runnableFuture =
+            executor.submit(
+                SentryWrapper.wrapRunnable {
+                    assertNotEquals(mainScopes, Sentry.getCurrentScopes())
+                    assertNotEquals(threadedScopes, Sentry.getCurrentScopes())
+                    "Result 1"
+                }
+            )
+
+        runnableFuture.get()
+
+        executor.submit {
+            assertNotEquals(mainScopes, Sentry.getCurrentScopes())
+            assertEquals(threadedScopes, Sentry.getCurrentScopes())
+        }.get()
+    }
 }